# Performance improvements for home page insight box around low grades