--- a/app/assets/javascripts/home/HomePage.test.js
+++ b/app/assets/javascripts/home/HomePage.test.js
@@ -10,7 +10,7 @@ import unsupportedLowGradesJson from '../../../../spec/javascripts/fixtures/home
 beforeEach(() => {
   fetchMock.restore();
   fetchMock.get('/home/feed_json?limit=20', homeFeedJson);
-  fetchMock.get('/home/unsupported_low_grades_json', unsupportedLowGradesJson);
+  fetchMock.get('/home/unsupported_low_grades_json?limit=100', unsupportedLowGradesJson);
 });
 
 it('renders without crashing', () => {