--- a/spec/javascripts/fixtures/home_unsupported_low_grades_json.js
+++ b/spec/javascripts/fixtures/home_unsupported_low_grades_json.js
@@ -1,4 +1,6 @@
 export default {
+  "limit": 100,
+  "total_count": 22,
   "assignments": [
     {
       "id": 3,