this builds on https://github.com/studentinsights/studentinsights/pull/1505

# Who is this PR for?
HS educators, particularly dept. heads or counselors who have access to many students

# What problem does this PR fix?
~25 second load time for the insight box about low grades, for dept. heads

# What does this PR do?
It optimizes the query for students with low grades but who haven't been discussed recently in NGE or 10GE.

# Screenshot (if adding a client-side feature)
No changes:
<img width="1016" alt="screen shot 2018-03-09 at 8 14 27 am" src="https://user-images.githubusercontent.com/1056957/37209104-e7346438-2371-11e8-9ec3-020c4be25b39.png">

# Checklists
+ [x] Author checked latest in IE - Home page
+ [x] Author included specs for new code