@kevinrobinson 2018-03-09
selfie ahead of the meeting this morning, @alexsoble can you review later today?